Fixes more logging, import errors
